+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Fix: NPE while adding "response_body_size" breadcrumb, when response body is null (#1884)
+
 ## 5.6.0
 
 * Fix: Change scope of servlet-api to compileOnly (#1880)